src/backend/requirements.txt
absl-py==1.4.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
boto3==1.24.96
botocore==1.27.96
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.32.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.8
django-after-response==0.2.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-environ==0.4.5
django-extensions==3.2.1
django-filter==22.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.13.1
django-webpack-loader==1.6.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
drf-nested-routers==0.93.4
drf-yasg==1.21.4
exceptiongroup==1.1.1
flatbuffers==23.5.8
gast==0.4.0
gensim==4.3.1
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
jax==0.4.9
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.1
ml-dtypes==0.1.0
msgpack==1.0.4
mysqlclient==2.1.1
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
Pillow==9.4.0
protobuf==4.23.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-logstash==0.4.8
python3-openid==3.2.0
pytz==2022.5
pyvi==0.1.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
sentry-sdk==1.11.0
simplejson==3.17.6
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.3.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2022.5
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.6
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
zipp==3.15.0
gunicorn==20.1.0
django-crontab==0.7.1
tensorrt==8.6.1
gunicorn==20.1.0
django-crontab==0.7.1
